refactor: move xtdb upsert orchestration
